% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combined_count.R
\name{combined_count}
\alias{combined_count}
\title{Count amount of intervals each day is in}
\usage{
combined_count(u = NULL, d = NULL, d_t = NULL, p_restrict = 0.95)
}
\arguments{
\item{u}{Result from GSADF with no drift or trend}

\item{d}{Result from GSADF with drift but no trend}

\item{d_t}{Result from GSADF with drift and trend}

\item{p_restrict}{Keep only observations with greater that p_restrict p value}
}
\value{
Number of intervals a day is observed in
}
\description{
This function counts the amount of intervals every day is in.
}
