% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DF_distribution.R
\name{DF_distribution}
\alias{DF_distribution}
\title{Simulate ADF distribution}
\usage{
DF_distribution(
  T_val = 100,
  seed = 12345,
  sd_rv = 1,
  y_0 = 1,
  n_monte = 10,
  T_vals_only = T,
  drift = F,
  trend = F
)
}
\arguments{
\item{T_val}{Sample size}

\item{seed}{Random seed}

\item{sd_rv}{Standard deviation for white noise}

\item{y_0}{Starting value}

\item{n_monte}{Number of simulations, should be in the thousands}

\item{T_vals_only}{TRUE if only the T values are wanted, if FALSE more things are returned}

\item{drift}{FALSE if no drift}

\item{trend}{FALSE if no trend}
}
\value{
ADF distribution for AR(1)
}
\description{
This function simulates the distribution for ADF
}
