% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discount.R
\name{discount}
\alias{discount}
\title{Discount prices}
\usage{
discount(p, r = 0.01)
}
\arguments{
\item{p}{Price}

\item{r}{Inflation rate}
}
\value{
discounted price
}
\description{
This function discounts prices to the first observation by the risk free rate(inflation)
}
