% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSADF.R
\name{GSADF}
\alias{GSADF}
\title{GSADF}
\usage{
GSADF(
  ticker,
  x = NULL,
  min_window = 30,
  step_length = 5,
  window_increase = 10,
  date_from = "1900-01-01",
  date_to = base::Sys.Date(),
  drift = F,
  trend = F,
  risk_free_rate = 0.01,
  own_df_distribution = NULL
)
}
\arguments{
\item{ticker}{Ticker code from yahoo finance of the asset}

\item{x}{vector of prices if no ticker code is available}

\item{min_window}{Minimum size of the window to test for unit root}

\item{step_length}{The size that the window will move for every step}

\item{window_increase}{How much the window size should increase every iteration}

\item{date_from}{Date from}

\item{date_to}{Date to}

\item{drift}{TRUE if drift, FALSE if no drift}

\item{trend}{TRUE if trend, FALSE if no trend}

\item{risk_free_rate}{Numeric value giving the yearly inflation}

\item{own_df_distribution}{If Null, will use simulated distributions in package, can input your own, generated from DF_distribution}
}
\value{
list with with "stock" containing the price of the asset and "result" containing
the result of the GSADF, meaning a a tibble containing interval lengths, estimates,
p values and more
}
\description{
This function is a implementation of GSADF as described by PSY
}
