% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_val_fun.R
\name{p_val_fun}
\alias{p_val_fun}
\title{Calculate the p value}
\usage{
p_val_fun(t_val, interval_length, distribution_tibble)
}
\arguments{
\item{t_val}{T value from lm call}

\item{interval_length}{The length of the interval}

\item{distribution_tibble}{The result from DF_distribution}
}
\value{
The p value as in P(x > X)
}
\description{
This function is used in GSADF to calculate the p value based on the simulated distribution
found from DF_distribution
}
