% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_GSADF.R
\name{plot_GSADF}
\alias{plot_GSADF}
\title{Plot the result of GSADF}
\usage{
plot_GSADF(
  u,
  d = NULL,
  d_t = NULL,
  p_restrict = 0.95,
  start_date_tq_get = "2020-01-01",
  image_name = NULL,
  valuta = "valuta",
  aktie = "aktie"
)
}
\arguments{
\item{u}{Result from GSADF with drift and trend = FALSE}

\item{d}{Result from GSADF with drif = TRUE and trend = FALSE}

\item{d_t}{Result from GSADF with drift = TRUE and trend = TRUE}

\item{p_restrict}{Keep only intervals with p value more than}

\item{start_date_tq_get}{The date from from your GSADF call}

\item{image_name}{Name of resulting pdf file, given in character}

\item{valuta}{Currency of asset, given in character}

\item{aktie}{Name of asset, given in character}
}
\value{
A plot if image_name = NULL, else create a pdf file with name image_name
}
\description{
This function takes the result of GSADF and returns a plot if the variable image_name is NULL
if it is not NULL then a pdf file will be created with the name given to image_name
}
