% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tabel_GSADF.R
\name{Tabel_GSADF}
\alias{Tabel_GSADF}
\title{Tabel GSADF}
\usage{
Tabel_GSADF(gsadf)
}
\arguments{
\item{gsadf}{Output from GSADF}
}
\value{
Summary of largest p-value
}
\description{
This function summarises the largest p-values of the result of GSADF
}
